backend/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==1.1.2
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==0.19.2
python-jose-cryptodome==1.3.2
six==1.12.0
SQLAlchemy==1.4.29
typed-ast==1.4.2
Werkzeug==0.15.6
wrapt==1.11.1
